src/back_up_read/requirements_ui.txt
streamlit
openai-whisper
tqdm
aiofiles
sqlmodel
pandas
dateparser
